pyproject.toml
[project]
name = "sherlock-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai>=1.100.2",
    "openai-agents>=0.2.8",
    "pydantic>=2.11.7",
    "python-nmap>=0.7.1",
    "requests>=2.32.5",
]
